+++ resources/js/scripts.js
@@ -80,6 +80,14 @@ document.addEventListener('alpine:init', () => {
         sortablejsObj: null,
         itemsIncrementInProgress: false,
 
+        init() {
+            $wire.watch('items', (value, oldValue) => {
+                if (value === null) {
+                    $wire.items = oldValue
+                }
+            })
+        },
+
         initDropzone() {
             _handleDropzone(this.$el, (e) => {
                 this.upload(e.dataTransfer.files)
@@ -124,6 +132,10 @@ document.addEventListener('alpine:init', () => {
         },
 
         itemHidden(itemId) {
+            if ($wire.items === null || ! $wire.items[itemId]) {
+                return false
+            }
+
             if (! $wire.autoSave && $wire.items[itemId].id !== null) {
                 return false
             }
@@ -132,7 +144,7 @@ document.addEventListener('alpine:init', () => {
         },
 
         visibleItemsNumber() {
-            return Object.keys($wire.items).filter(itemId => ! this.itemHidden(itemId)).length
+            return Object.keys($wire.items ?? {}).filter(itemId => ! this.itemHidden(itemId)).length
         },
 
         maxFilesNumberReached() {
@@ -216,6 +228,12 @@ document.addEventListener('alpine:init', () => {
         deleteTimerInterval: null,
 
         init() {
+            $wire.watch('itemData', (value, oldValue) => {
+                if (value === null) {
+                    $wire.itemData = oldValue
+                }
+            })
+
             if ($wire.uploadFromGroupAtIndex !== null) {
                 this.upload(this.filesFromGroup[$wire.uploadFromGroupAtIndex])
             }